# Conflicts:
#	Assets/Scripts/ECS/EcsStartup.cs
#	Assets/Scripts/ECS/Systems/Input/CheckInvokeAbilitySystem.cs
#	Assets/Scripts/ECS/Systems/Unit/InitPlayerSystems.cs

+++ Assets/Scripts/Ability/Ability.cs
@@ -6,7 +6,7 @@ using UnityEngine.Tilemaps;
 public class Ability
 {
     public float Damage;
-    public float Cost;
+    public int Cost;
     public TileBase TileView;
     [SerializeReference] public AttackZoneConfig AttackZoneConfig;
     [SerializeReference] public AttackAreaConfig AttackAreaConfig;
